+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - error handling for getting teams and channels
 - error handling for some missing parameters
 - fixing the package layout, also including abstract.py
+- handle invalid tokens
 
 ## 0.1 (2018-06-18)
 - initial version